update cross validation and stacking
