+++ scripts/diff_vocab.py
@@ -1,4 +1,3 @@
-# -*- encoding:utf-8 -*-
 import sys
 import os
 import argparse
